4 h

+++ base/views.py
@@ -1,6 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from .models import Room , Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Massage
 from .forms import RoomForm
 
 
@@ -14,11 +20,65 @@ from .forms import RoomForm
 
 
 def loginPage(request):
-    context = {}
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        print(username, password)
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Username does not exist')
+
+            # if user.check_password(password):
+            #     print('password correct')
+            #     login(request, user)
+            #     return redirect('home')
+            # else:
+            #     print('password incorrect')
+            #     messages.info(request, 'Username OR password is incorrect')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password is incorrect')
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)  # commit=False means not save in database
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'base/login_register.html', {'form': form, 'page': page})
+
+
 def home(request):
-    q = request.GET.get('q') if request.GET.get('q') is not None else '' #search
+    q = request.GET.get('q') if request.GET.get('q') is not None else ''  # search
     rooms = Room.objects.all().filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
@@ -27,17 +87,40 @@ def home(request):
 
     topics = Topic.objects.all()
     room_count = rooms.count()
+    room_message = Massage.objects.filter(Q(room__topic__name__icontains=q))
 
-    context = {'rooms': rooms , 'topics':topics , 'room_count':room_count}
+    context = {'rooms': rooms, 'topics': topics,
+               'room_count': room_count, 'room_message': room_message}
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.massage_set.all()
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Massage.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     return render(request, 'base/room.html', context)
 
 
+def userProfile(request, pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.massage_set.all()
+
+    context = {'user': user, 'rooms': rooms, 'room_messages': room_messages}
+    return render(request, 'base/profile.html', context)
+
+
+@login_required(login_url='login')  # if user is not login then redirect to login page
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -50,9 +133,14 @@ def createRoom(request):
     return render(request, 'base/room_from.html', context)
 
 
+@login_required(login_url='login')  # if user is not login then redirect to login page
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
 
         form = RoomForm(request.POST, instance=room)
@@ -63,10 +151,29 @@ def updateRoom(request, pk):
     return render(request, 'base/room_from.html', context)
 
 
+@login_required(login_url='login')  # if user is not login then redirect to login page
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     return render(request, 'base/delete.html', {'obj': room})
+
+
+@login_required(login_url='login')  # if user is not login then redirect to login page
+def deleteMessage(request, pk):
+    message = Massage.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+
+    return render(request, 'base/delete.html', {'obj': message})
